actions: Split out zerver.lib.user_message.
